// frontend/src/app/api/model/dashboard.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DashboardHabits } from './dashboardHabits';
import { DashboardTasks } from './dashboardTasks';
import { DashboardPlans } from './dashboardPlans';
import { DashboardLearning } from './dashboardLearning';


export interface Dashboard { 
    greetingName: string;
    today: string;
    tasks: DashboardTasks;
    habits: DashboardHabits;
    plans: DashboardPlans;
    learning: DashboardLearning;
}